pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dragent"
version = "0.1.0"
description = "Agentic AI for Residential Demand Response"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain==0.1.0",
    "langchain-openai",
    "langchain-core",
    "cvxpy>=1.4.0",
    "numpy>=1.24.0",
    "openai>=1.56.1",
    "python-dotenv>=1.0.0",
    "ecos>=2.0.0",
    "gradio>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dragent*"]
